requirements.txt
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.2
pyparsing==2.4.7
pytesseract==0.3.7
python-dateutil==2.8.1
six==1.15.0
xlib==0.21
